web: fix `WebCommandList` links
