--- client/web/src/components/WebCommandListPopoverButton.module.scss
@@ -1,8 +1,0 @@
-.button {
-    color: var(--icon-color);
-
-    &:hover {
-        // Required to overwrite .btn:hover styles with greater specificity.
-        color: var(--body-color) !important;
-    }
-}
