+++ client/web/src/components/WebCommandListPopoverButton/WebCommandListPopoverButton.module.scss
@@ -1,0 +1,19 @@
+.button {
+    color: var(--icon-color);
+
+    &:hover {
+        // Required to overwrite .btn:hover styles with greater specificity.
+        color: var(--body-color) !important;
+    }
+}
+
+.action-item {
+    --link-color: var(--body-color);
+    --link-hover-color: var(--body-color);
+
+    display: block;
+    text-align: left;
+
+    /* Overwrite `ButtonLink` style */
+    text-decoration: none !important;
+}
