+++ client/web/src/components/WebCommandListPopoverButton/index.ts
@@ -1,0 +1,1 @@
+export * from './WebCommandListPopoverButton'
